add refresh token

+++ src/main/java/com/example/tinywiny/service/ProductService.java
@@ -1,11 +1,17 @@
 package com.example.tinywiny.service;
 
-import com.example.tinywiny.converter.ProductConverter;
-import com.example.tinywiny.converter.TypeProductConverter;
 import com.example.tinywiny.dto.ProductDto;
+import com.example.tinywiny.dto.ProductDtoWithImage;
 import com.example.tinywiny.dto.TypeProduct;
+import com.example.tinywiny.model.Image;
 import com.example.tinywiny.model.Product;
+import com.example.tinywiny.model.ProductInBucket;
+import com.example.tinywiny.model.ProductInOrder;
+import com.example.tinywiny.repository.ImageRepository;
+import com.example.tinywiny.repository.ProductInBucketRepository;
+import com.example.tinywiny.repository.ProductInOrderRepository;
 import com.example.tinywiny.repository.ProductRepository;
+import com.example.tinywiny.service.amazonService.AwsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Page;
@@ -15,6 +21,9 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,17 +32,15 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ProductService {
   private final ProductRepository productRepository;
-  private final ProductConverter converter;
+  private final ProductInBucketRepository productInBucketRepository;
+  private final ProductInOrderRepository productInOrderRepository;
   private final TypeProductService typeProductService;
-  private final TypeProductConverter typeProductConverter;
-
+  private final ImageRepository imageRepository;
+  private final AwsService storageService;
 
   @Transactional
   @Modifying
   public Product save(Product product) {
-    if (productRepository.findProductByProductName(product.getProductName()).isPresent()) {
-      throw new RuntimeException("User already exists");
-    }
     return productRepository.save(product);
   }
 
@@ -50,9 +57,6 @@ public class ProductService {
   }
 
   private Product prepareProductForUpdate(ProductDto productDto, Product product) {
-    if (productDto.getProductName() != null) {
-      product.setProductName(productDto.getProductName());
-    }
     if (productDto.getPrice() != 0) {
       product.setPrice(productDto.getPrice());
     }
@@ -99,17 +103,49 @@ public class ProductService {
     }
     return product.get();
   }
-/*
-  public int findProductPrice(Long productId) {
-    return productRepository.findProductPrice(productId);
-  }*/
 
   public void deleteProduct(Long productId) {
-    productRepository.deleteProductByProductId(productId);
+    ProductInBucket productInBucket = productInBucketRepository.findProductInBucketById(productId);
+    ProductInOrder productInOrder = productInOrderRepository.
+        findProductInOrderByProduct(findProductByProductId(productId));
+    if (productInBucket == null && productInOrder == null)
+      productRepository.deleteProductByProductId(productId);
   }
-
 
   public List<Product> findAllProducts() {
     return productRepository.findAll();
+  }
+
+  public ProductDtoWithImage productConverter(ProductDto productDto, URI image) {
+    ProductDtoWithImage productDtoWithImage = new ProductDtoWithImage();
+    productDtoWithImage.setProductId(productDto.getProductId());
+    productDtoWithImage.setCountInStock(productDto.getCountInStock());
+    productDtoWithImage.setProductName(productDto.getProductName());
+    productDtoWithImage.setPrice(productDto.getPrice());
+    productDtoWithImage.setDescription(productDto.getDescription());
+    productDtoWithImage.setIdType(productDto.getIdType());
+    productDtoWithImage.setImage(image);
+    return productDtoWithImage;
+  }
+
+  public List<ProductDtoWithImage> productConverter(List<ProductDto> productDto) throws URISyntaxException {
+    List<ProductDtoWithImage> products = new ArrayList<>();
+    for (ProductDto prod : productDto) {
+      URI imageUrl = null;
+      Image image = imageRepository.findImageByProductId(prod.getProductId());
+      if (image != null) {
+        imageUrl = storageService.getImagePath(image.getImageName());
+      }
+      ProductDtoWithImage productDtoWithImage = new ProductDtoWithImage();
+      productDtoWithImage.setProductId(prod.getProductId());
+      productDtoWithImage.setCountInStock(prod.getCountInStock());
+      productDtoWithImage.setProductName(prod.getProductName());
+      productDtoWithImage.setPrice(prod.getPrice());
+      productDtoWithImage.setDescription(prod.getDescription());
+      productDtoWithImage.setIdType(prod.getIdType());
+      productDtoWithImage.setImage(imageUrl);
+      products.add(productDtoWithImage);
+    }
+    return products;
   }
 }
